Fix dark mode not starting with the browser saved variable
