+++ src/frontend/src/index.tsx
@@ -12,21 +12,15 @@ import { RouterProvider } from "react-router-dom";
 import LoadingComponent from "./components/loadingComponent";
 import router from "./routes";
 import "./style/classes.css";
+import App from "./App";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement,
 );
+
 root.render(
   <ContextWrapper>
-    <Suspense
-      fallback={
-        <div className="loading-page-panel">
-          <LoadingComponent remSize={50} />
-        </div>
-      }
-    >
-      <RouterProvider router={router} />
-    </Suspense>
+    <App />
   </ContextWrapper>,
 );
 reportWebVitals();
